\name{adjustment}
\alias{adjustment}
\title{Access the multiple comparison adjustment method from the regressResult or interactionResult class}
\description{
  Access the multiple comparison adjustment method from the regressResult class or interactionResult class
}
\usage{
adjustment(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class }
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }
