\name{arraytools-package}
\alias{arraytools-package}
\alias{arraytools}
\docType{package}
\title{
Tools for working with arrays
}
\description{
Provides tools for working with arrays.
}
\details{
\tabular{ll}{
Package: \tab arraytools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-20\cr
License: \tab MIT\cr
}

}
\author{
Peyton McCullough <peyton@peytonmccullough.com>
}
