\name{eSetExample}
\alias{eSetExample}
\docType{data}
\title{An ExpressionSet example}
\description{
  An ExpressionSet example
}
\usage{data(eSetExample)}
\format{
  The format is:
  Formal class 'ExpressionSet' [package "Biobase"] with 6 slots
}
\examples{
data(eSetExample)
}
\keyword{datasets}
