\name{exprsExample}
\alias{exprsExample}
\docType{data}
\title{a data.frame contains expression data}
\description{
  a data.frame contains expression data
}
\usage{data(exprsExample)}
\format{
  A data frame with 1000 observations on the following 17 variables.
  \describe{
    \item{\code{probeset_id}}{a numeric vector}
    \item{\code{H1.CEL}}{a numeric vector}
    \item{\code{H2.CEL}}{a numeric vector}
    \item{\code{H3.CEL}}{a numeric vector}
    \item{\code{H4.CEL}}{a numeric vector}
    \item{\code{H5.CEL}}{a numeric vector}
    \item{\code{H6.CEL}}{a numeric vector}
    \item{\code{H7.CEL}}{a numeric vector}
    \item{\code{H8.CEL}}{a numeric vector}
    \item{\code{H9.CEL}}{a numeric vector}
    \item{\code{H10.CEL}}{a numeric vector}
    \item{\code{H11.CEL}}{a numeric vector}
    \item{\code{H12.CEL}}{a numeric vector}
    \item{\code{H13.CEL}}{a numeric vector}
    \item{\code{H14.CEL}}{a numeric vector}
    \item{\code{H15.CEL}}{a numeric vector}
    \item{\code{H16.CEL}}{a numeric vector}
  }
}
\examples{
data(exprsExample)
}
\keyword{datasets}
