\name{getAdjP}
\alias{getAdjP}
\title{ access the adjPVal slot from regressResult or interactionResult class}
\description{
  access the adjPVal slot from regressResult or interactionResult class
}
\usage{
getAdjP(object)
}
\arguments{
  \item{object}{a regressResult class or interactionResult class}
}
\value{
  a numeric vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }
