\name{getAnnotation}
\alias{getAnnotation}
\title{access the annotation slot from the regressResult or interactionResult slot}
\description{
  access the annotation slot from the regressResult or interactionResult slot
}
\usage{
getAnnotation(object)
}
\arguments{
  \item{object}{a regressResult class or interactionResult class}
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }

