\name{getCompare1}
\alias{getCompare1}
\title{Access the Compare1 slot from the contrastMatrix }
\description{
  Access the Compare1 slot from the contrastMatrix 
}
\usage{
getCompare1(object)
}
\arguments{
  \item{object}{a contrastMatrix class}
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{contrastMatrix}} }
