\name{getCompare2}
\alias{getCompare2}
\title{Access the compare2 slot from the contrastMatrix class }
\description{
  Access the compare2 slot from the contrastMatrix class
}
\usage{
getCompare2(object)
}
\arguments{
  \item{object}{a contrastMatrix class }
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{contrastMatrix}} }