\name{getContrast}
\alias{getContrast}
\title{Access the contrast matrix from the contrastMatrix class}
\description{
  Access the contrast matrix from the contrastMatrix class
}
\usage{
getContrast(object)
}
\arguments{
  \item{object}{a contrastMatrix class}
}
\value{
  a numeric matrix
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{contrastMatrix}} }