\name{getCovariates}
\alias{getCovariates}
\title{Accessing the covariates from the designMatrix class}
\description{
  Accessing the covariates from the designMatrix class
}
\usage{
getCovariates(object)
}
\arguments{
  \item{object}{a designMatrix class}
}
\value{
  a character vector containing \code{covariates}
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{designMatrix}} }

