\name{getDesign}
\alias{getDesign}
\title{Access the design matrix from the designMatrix class }
\description{
  Access the design matrix from the designMatrix class 
}
\usage{
getDesign(object)
}
\arguments{
  \item{object}{a \code{designMatrix} class }
}
\value{
  a matrix containing the \code{designMatrix}
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{designMatrix}} }
