\name{getF}
\alias{getF}
\title{access the foldChange slot from regressionResult or interactionResult class}
\description{
  access the \code{foldChange} slot from regressionResult or interactionResult class
}
\usage{
getF(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class}
}
\value{
  a numeric vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }
