\name{getFC}
\alias{getFC}
\title{Access the foldChange slot from the regressResult or interactionResult class}
\description{
  Access the foldChange slot from the regressResult or interactionResult class
}
\usage{
getFC(object)
}
\arguments{
  \item{object}{a regressResult class or interactionResult class }
}
\value{
  a numeric vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }