\name{getFCCutoff}
\alias{getFCCutoff}
\title{Access the significantFCCutoff slot from the regressResult or interactionResult class}
\description{
  Access the significantFCCutoff slot from the regressResult or interactionResult class
}
\usage{
getFCCutoff(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class}
}
\value{
  a numeric vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }

