\name{getFilterMethod}
\alias{getFilterMethod}
\title{Access the filterMethod slot from the regressResult or interactionResult class}
\description{
  Access the filterMethod slot from the regressResult or interactionResult class
}
\usage{
getFilterMethod(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class }
}
\value{
  a list
}
\references{ ~put references to the literature/web site here ~ }
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }