\name{getID}
\alias{getID}
\title{access the ID slot from the regressResult or interactionResult class }
\description{
  access the ID slot from the regressResult or interactionResult class 
}
\usage{
getID(object)
}
\arguments{
  \item{object}{a regressResult class or interactionResult class}
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }