\name{getIndex}
\alias{getIndex}
\title{ Access the SignificantIndex slot from the regressResult or interactionResult class }
\description{
  Access the SignificantIndex slot from the regressResult or interactionResult class
}
\usage{
getIndex(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class}
}
\value{
  a logical vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }