\name{getInteraction}
\alias{getInteraction}
\title{Access the interaction slot from the contrastMatrix class}
\description{
  Access the interaction slot from the contrastMatrix class
}
\usage{
getInteraction(object)
}
\arguments{
  \item{object}{a contrastMatrix class}
}
\value{
  a logical vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{contrastMatrix}} }