\name{getIntIndex}
\alias{getIntIndex}
\title{Access the IntIndex slot from the designMatrix class}
\description{
  Access the IntIndex slot from the designMatrix class
}
\usage{
getIntIndex(object)
}
\arguments{
  \item{object}{an \code{designMatrix} class}
}
\value{
  a mumeric vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{designMatrix}} }