\name{getLength}
\alias{getLength}
\title{Calculate the Length of interactionResult class}
\description{
  Calculate the Length of interactionResult class
}
\usage{
getLength(object)
}
\arguments{
  \item{object}{an interactionResult class}
}
\value{
  a numeric value
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{interactionResult}} }