\name{getLevel}
\alias{getLevel}
\title{Access the level slot from the contrastMatrix class}
\description{
  Access the level slot from the contrastMatrix class
}
\usage{
getLevel(object)
}
\arguments{
  \item{object}{a contrastMatrix class}
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{contrastMatrix}} }
