\name{getP}
\alias{getP}
\title{Access the pValue slot from regressResult or interactionResult class}
\description{
  Access the pValue slot from regressResult or interactionResult class
}
\usage{
getP(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class }
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }
