\name{getPCutoff}
\alias{getPCutoff}
\title{Access the significantPvalueCutoff slot from regressResult or interactionResult class}
\description{
  Access the significantPvalueCutoff slot from regressResult or interactionResult class
}
\usage{
getPCutoff(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class}
}
\value{
  a numeric vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }