\name{getTarget}
\alias{getTarget}
\title{Access the target slots from the designMatrix class}
\description{
  Access the target slots from the designMatrix class
}
\usage{
getTarget(object)
}
\arguments{
  \item{object}{a \code{designMatrix} class}
}
\value{
   a data frame contains the target file
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{designMatrix}} }

