\name{hugene10stCONTROL}
\alias{hugene10stCONTROL}
\docType{data}
\title{hugene10stCONTROL}
\description{
  It is used to remove "normgene" and "control" genes for hugene10st array
  in the \code{preProcessGeneST} function.  It is not intended to be used
  by the user.
}
\usage{data(hugene10stCONTROL)}
\format{
  A data frame with 4201 observations on the following 2 variables.
}

