\name{mogene10stCONTROL}
\alias{mogene10stCONTROL}
\docType{data}
\title{mogene10stCONTROL }
\description{
  It is used to remove "normgene" and "control" genes for mogene10st array
  in the \code{preProcessGeneST} function.  It is not intended to be used
  by the user.
}
\usage{data(mogene10stCONTROL)}
\format{
  A data frame with 6613 observations on the following 2 variables.
}
