\name{output.cls}
\alias{output.cls}
\title{Create *.CLS file for GSEA analysis}
\description{
  Create *.CLS file for GSEA analysis
}
\usage{
output.cls(target, variable, filename = "phenotype")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{target}{pheno Data file}
  \item{variable}{variable of interest}
  \item{filename}{file name}
}
\value{
  create a *.CLS file
}
\references{\code{http://www.broad.mit.edu/gsea/}}
\author{Xiwei, Wu, Arthur Li}
\seealso{\code{\link{output.gct}}, \code{\link{createGSEAFiles}}}
