\name{output.gct}
\alias{output.gct}
\title{Create an *.GCT file for GSEA analysis}
\description{
  Create an *.GCT file for GSEA analysis
}
\usage{
output.gct(normal, filename = "probe")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{normal}{an ExpressionSet}
  \item{filename}{file name}
}
\value{
  create an *.GCT file 
}
\references{\code{http://www.broad.mit.edu/gsea/}}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\link{output.cls}}, \code{\link{createGSEAFiles}}}


