\name{output.ing}
\alias{output.ing}
\title{Create an Ingenuity File for Ingenuity Analysis}
\description{
  Create an Ingenuity File for Ingenuity Analysis
}
\usage{
output.ing(allfile, eSet, filename = "IngenuityFile")
}
\arguments{
  \item{allfile}{a list of regressResult class}
  \item{eSet}{an ExpressionSet}
  \item{filename}{file name }
}
\value{
  create an txt file for Ingenuity Analysis
}
\references{\code{http://www.ingenuity.com/}}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\link{createIngenuityFile}}}
