\name{pDataExample}
\alias{pDataExample}
\docType{data}
\title{a phenoData example}
\description{
  a data frame contains the phenotype data
}
\usage{data(pDataExample)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Treatment}}{a character vector}
    \item{\code{Group}}{a character vector}
  }
}
\examples{
data(pDataExample)
}
\keyword{datasets}
