\name{qaGeneST}
\alias{qaGeneST}
\title{Creating Quality Assessment Report for Gene ST Array}
\description{
  Creating Quality Assessment Report for Gene ST Array in HTML file
}
\usage{
qaGeneST(object, parameters, QC, mydir = getwd(), outputFile = "QA.html")
}
\arguments{
  \item{object}{an \code{ExpressionSet}}
  \item{parameters}{The names of the variables to be included in the report }
  \item{QC}{The QC report generated from Affymetrix Expression Console}
  \item{mydir}{The name of the directory containing the report}
  \item{outputFile}{The name of the outputfile.  Make sure write ".html"}
}

\details{
This function creates quality control report in an HTML file that contains a 
set of 8 assessment figures.  

Figure1: The intensity distributio Plot. The raw intensity should be similar 
across all chips

Figure2: The Mean Signal Plot.  The mean signal of each group should be consistant
across the samples.  The positive control should be higher than the negative controls.

Figure3: BAC SPIKE plot. The mean signal of each group should be consistant
across the samples. The signal for BioB should be the lowest, follows by
BioC, BioD, and CreX (the highest).

Figure4: POLYA SPIKE plot. The mean signal of each group should be consistant
across the samples. The signal for Lys should be the lowest, follows by Thr, Phe, and Dap.

Figure5: POS VS NEG AUC plot. Pos vs neg auc is the area under the curve (AUC) for a 
receiver operating characteristic (ROC) plot comparing signal values for the positive 
controls to the negative controls.  In practice the expected value for this metric
is tissue type specific and may be sensitive to the quality of the RNA sample.
Values between 0.80 and 0.90 are typical.

Figure6: MAD RESIDUAL MEAN plot. A measure of how well or poor all of the probes on a 
given chip fit the RMA or PLIER model. An unusually high mean absolute
deviation of the residuals from the median suggests problematic data for that
chip.

Figure7: RLE MEAN plot. This metric is generated by taking the
signal estimate for a given probeset on a given chip and calculating the
difference in log base 2 from the median signal value of that probeset over all the
chips. When just the replicates are analyzed together the mean absolute RLE should
be consistently low, reflecting the low biological variability of the replicates.

Figure8: Hierarchical Clustering of Samples .  Samples will be grouped using 
hierarchical clustering and principal component analysis (PCA). If the sample 
preparation steps introduced bigger variation than biological variation, 
treatment groups will be mixed up in the plot. This could also happen when the 
samples between groups were mixed up accidentally when the samples were prepared. 
We acknowledge that clinical samples are harder to collect and sometimes impossible 
to control. Therefore, sample QC criteria will be much looser when dealing with 
clinical samples.        
}
\value{
  no value is returned
}
\references{
  \url{http://www.affymetrix.com/support/technical/whitepapers/exon_gene_arrays_qa_whitepaper.pdf}}

\author{Xiwei Wu, Arthur Li}
\examples{
data(eSetExample)
logdata <- preProcessGeneST(eSetExample)
data(QC)
\dontrun{qaGeneST(logdata,  c("Treatment", "Group"), QC)}
}
