\name{regressionMethod}
\alias{regressionMethod}
\title{Access the regressionMethod slot from the regressResult or interactionResult class}
\description{
  Access the regressionMethod slot from the regressResult or interactionResult class
}
\usage{
regressionMethod(object)
}
\arguments{
  \item{object}{a regressResult or interactionResult class}
}
\value{
  a character vector
}
\author{Xiwei Wu, Arthur Li}
\seealso{\code{\linkS4class{regressResult}} \code{\linkS4class{interactionResult}} }